+++ test/__tests__/cameras.test.js
@@ -149,8 +149,6 @@ describe('DELETE /api/cameras/:name', () => {
       .delete('/api/cameras/Test Camera 3')
       .auth(auth.body.access_token, { type: 'bearer' });
 
-    console.log(response.error);
-
     expect(response.statusCode).toBe(204);
   });
 });