+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+# v1.0.7 - 2022-01-15
+
+## Other Changes
+- Videoanalysis: Reduced the dwell time from 90s to 60s
+- Videoanalysis: Removed `-hwaccel` from FFMPEG parameters
+- Videoanalysis: Added `pixel/color difference` slider to be able to adjust the video analysis even more precisely within ui
+- Prebuffering: Added `-preset:v ultrafast` if `forcePrebuffering` is enabled
+- Added the possibility to control the motion sensor (OFF state) via the camera instead of via motionTimeout (set `"motionTimeout": 0`, the camera must be able to send a `OFF` message e.g. via MQTT or Videoanalysis)
+- More translation added
+- Minor improvements (camera.ui)
+
+## Bugfixes
+- Fixed an isue where crashing FFmpeg did not display an error message in the log
+- Fixed an issue where the dwell time could start before the motion handler was initialized
+- Fixed an issue where the restart timer for prebuffering and videoanalysis were calculated wrong
+- Minor Bugfixes
+
 # v1.0.6 - 2022-01-14
 
 ## Other Changes